moves logo files to assts foldere
