+++ client/assets/scranscanner-icon-dark.png
