+++ client/assets/scranscanner-icon-pink.png
