+++ client/assets/scranscanner-icon-white.png
